move peer manager to p2p_service
