+++ crates/services/p2p/src/lib.rs
@@ -5,6 +5,7 @@ mod discovery;
 mod gossipsub;
 mod heartbeat;
 mod p2p_service;
+mod peer_report;
 mod peer_manager;
 pub mod ports;
 mod request_response;
