+++ crates/services/p2p/src/peer_manager.rs
@@ -1,11 +1,3 @@
-use crate::{
-    config::Config,
-    gossipsub_config::PeerScoreConfig,
-    heartbeat::{
-        Heartbeat,
-        HeartbeatEvent,
-    },
-};
 use fuel_core_types::{
     blockchain::primitives::BlockHeight,
     services::p2p::peer_reputation::{
@@ -16,565 +8,30 @@ use fuel_core_types::{
     },
 };
 use libp2p::{
-    core::{
-        connection::ConnectionId,
-        either::EitherOutput,
-    },
-    identify::{
-        Behaviour as Identify,
-        Config as IdentifyConfig,
-        Event as IdentifyEvent,
-        Info as IdentifyInfo,
-    },
-    swarm::{
-        derive_prelude::{
-            ConnectionClosed,
-            ConnectionEstablished,
-            DialFailure,
-            FromSwarm,
-            ListenFailure,
-        },
-        ConnectionHandler,
-        IntoConnectionHandler,
-        IntoConnectionHandlerSelect,
-        NetworkBehaviour,
-        NetworkBehaviourAction,
-        PollParameters,
-    },
     Multiaddr,
     PeerId,
 };
 use rand::seq::IteratorRandom;
-
 use std::{
     collections::{
         HashMap,
         HashSet,
-        VecDeque,
     },
     sync::{
         Arc,
         RwLock,
     },
-    task::{
-        Context,
-        Poll,
-    },
-    time::{
-        Duration,
-        Instant,
-    },
+    time::Duration,
 };
-use tokio::time::{
-    self,
-    Interval,
-};
+use tokio::time::Instant;
+use tracing::debug;
 
-use tracing::{
-    debug,
-    info,
-};
+use crate::gossipsub_config::PeerScoreConfig;
 
-/// Maximum amount of peer's addresses that we are ready to store per peer
-const MAX_IDENTIFY_ADDRESSES: usize = 10;
-const HEALTH_CHECK_INTERVAL_IN_SECONDS: u64 = 10;
-const REPUTATION_DECAY_INTERVAL_IN_SECONDS: u64 = 1;
-
-/// Events emitted by PeerInfoBehaviour
-#[derive(Debug, Clone)]
-pub enum PeerInfoEvent {
-    PeerConnected(PeerId),
-    PeerDisconnected {
-        peer_id: PeerId,
-        should_reconnect: bool,
-    },
-    TooManyPeers {
-        peer_to_disconnect: PeerId,
-    },
-    BanPeer {
-        peer_id: PeerId,
-    },
-    ReconnectToPeer(PeerId),
-    PeerIdentified {
-        peer_id: PeerId,
-        addresses: Vec<Multiaddr>,
-    },
-    PeerInfoUpdated {
-        peer_id: PeerId,
-        block_height: BlockHeight,
-    },
-}
-
-// `Behaviour` that holds info about peers
-pub struct PeerManagerBehaviour {
-    heartbeat: Heartbeat,
-    identify: Identify,
-    peer_manager: PeerManager,
-    // regulary checks if reserved nodes are connected
-    health_check: Interval,
-    decay_interval: Interval,
-}
-
-impl PeerManagerBehaviour {
-    pub(crate) fn new(
-        config: &Config,
-        connection_state: Arc<RwLock<ConnectionState>>,
-        peer_score_config: PeerScoreConfig,
-    ) -> Self {
-        let identify = {
-            let identify_config =
-                IdentifyConfig::new("/fuel/1.0".to_string(), config.keypair.public());
-            if let Some(interval) = config.identify_interval {
-                Identify::new(identify_config.with_interval(interval))
-            } else {
-                Identify::new(identify_config)
-            }
-        };
-
-        let heartbeat =
-            Heartbeat::new(config.heartbeat_config.clone(), BlockHeight::default());
-
-        let reserved_peers: HashSet<PeerId> = config
-            .reserved_nodes
-            .iter()
-            .filter_map(PeerId::try_from_multiaddr)
-            .collect();
-
-        let peer_manager = PeerManager::new(
-            peer_score_config,
-            reserved_peers,
-            connection_state,
-            config.max_peers_connected as usize,
-        );
-
-        Self {
-            heartbeat,
-            identify,
-            peer_manager,
-            health_check: time::interval(Duration::from_secs(
-                HEALTH_CHECK_INTERVAL_IN_SECONDS,
-            )),
-            decay_interval: time::interval(Duration::from_secs(
-                REPUTATION_DECAY_INTERVAL_IN_SECONDS,
-            )),
-        }
-    }
-
-    pub fn total_peers_connected(&self) -> usize {
-        self.peer_manager.total_peers_connected()
-    }
-
-    /// returns an iterator over the connected peers
-    pub fn get_peers_ids(&self) -> impl Iterator<Item = &PeerId> {
-        self.peer_manager.get_peers_ids()
-    }
-
-    pub fn get_peer_info(&self, peer_id: &PeerId) -> Option<&PeerInfo> {
-        self.peer_manager.get_peer_info(peer_id)
-    }
-
-    pub fn insert_peer_addresses(&mut self, peer_id: &PeerId, addresses: Vec<Multiaddr>) {
-        self.peer_manager
-            .insert_peer_info(peer_id, PeerInfoInsert::Addresses(addresses));
-    }
-
-    pub fn update_block_height(&mut self, block_height: BlockHeight) {
-        self.heartbeat.update_block_height(block_height);
-    }
-
-    /// Find a peer that is holding the given block height.
-    pub fn get_peer_id_with_height(&self, height: &BlockHeight) -> Option<PeerId> {
-        self.peer_manager.get_peer_id_with_height(height)
-    }
-
-    pub fn report_peer(
-        &mut self,
-        peer_id: PeerId,
-        peer_score: PeerScore,
-        reporting_service: &str,
-    ) -> Option<PeerScore> {
-        if let Some(latest_peer_score) = self
-            .peer_manager
-            .update_peer_score_with(peer_id, peer_score)
-        {
-            info!(target: "fuel-p2p", "{reporting_service} updated {peer_id} with new score {latest_peer_score}");
-
-            Some(latest_peer_score)
-        } else {
-            None
-        }
-    }
-
-    pub fn report_gossip_score(&mut self, peer_id: &PeerId, gossip_score: f64) {
-        if self
-            .peer_manager
-            .non_reserved_connected_peers
-            .contains_key(peer_id)
-            && gossip_score
-                < self
-                    .peer_manager
-                    .peer_score_config
-                    .get_min_gossipsub_score()
-        {
-            self.peer_manager
-                .pending_events
-                .push_front(PeerInfoEvent::BanPeer { peer_id: *peer_id })
-        }
-    }
-}
-
-impl NetworkBehaviour for PeerManagerBehaviour {
-    type ConnectionHandler = IntoConnectionHandlerSelect<
-        <Heartbeat as NetworkBehaviour>::ConnectionHandler,
-        <Identify as NetworkBehaviour>::ConnectionHandler,
-    >;
-    type OutEvent = PeerInfoEvent;
-
-    fn new_handler(&mut self) -> Self::ConnectionHandler {
-        IntoConnectionHandler::select(
-            self.heartbeat.new_handler(),
-            self.identify.new_handler(),
-        )
-    }
-
-    fn addresses_of_peer(&mut self, peer_id: &PeerId) -> Vec<Multiaddr> {
-        self.identify.addresses_of_peer(peer_id)
-    }
-
-    fn on_swarm_event(&mut self, event: FromSwarm<Self::ConnectionHandler>) {
-        match event {
-            FromSwarm::ConnectionEstablished(connection_established) => {
-                let ConnectionEstablished {
-                    peer_id,
-                    other_established,
-                    ..
-                } = connection_established;
-
-                self.heartbeat
-                    .on_swarm_event(FromSwarm::ConnectionEstablished(
-                        connection_established,
-                    ));
-                self.identify
-                    .on_swarm_event(FromSwarm::ConnectionEstablished(
-                        connection_established,
-                    ));
-
-                if other_established == 0 {
-                    // this is the first connection to a given Peer
-                    self.peer_manager.handle_initial_connection(peer_id);
-                }
-
-                let addresses = self.addresses_of_peer(&peer_id);
-                self.insert_peer_addresses(&peer_id, addresses);
-            }
-            FromSwarm::ConnectionClosed(connection_closed) => {
-                let ConnectionClosed {
-                    remaining_established,
-                    peer_id,
-                    connection_id,
-                    endpoint,
-                    ..
-                } = connection_closed;
-
-                let (ping_handler, identity_handler) =
-                    connection_closed.handler.into_inner();
-
-                let ping_event = ConnectionClosed {
-                    handler: ping_handler,
-                    peer_id,
-                    connection_id,
-                    endpoint,
-                    remaining_established,
-                };
-                self.heartbeat
-                    .on_swarm_event(FromSwarm::ConnectionClosed(ping_event));
-
-                let identify_event = ConnectionClosed {
-                    handler: identity_handler,
-                    peer_id,
-                    connection_id,
-                    endpoint,
-                    remaining_established,
-                };
-
-                self.identify
-                    .on_swarm_event(FromSwarm::ConnectionClosed(identify_event));
-
-                if remaining_established == 0 {
-                    // this was the last connection to a given Peer
-                    self.peer_manager.handle_peer_disconnect(peer_id);
-                }
-            }
-            FromSwarm::AddressChange(e) => {
-                self.heartbeat.on_swarm_event(FromSwarm::AddressChange(e));
-                self.identify.on_swarm_event(FromSwarm::AddressChange(e));
-            }
-            FromSwarm::DialFailure(e) => {
-                let (ping_handler, identity_handler) = e.handler.into_inner();
-                let ping_event = DialFailure {
-                    peer_id: e.peer_id,
-                    handler: ping_handler,
-                    error: e.error,
-                };
-                let identity_event = DialFailure {
-                    peer_id: e.peer_id,
-                    handler: identity_handler,
-                    error: e.error,
-                };
-                self.heartbeat
-                    .on_swarm_event(FromSwarm::DialFailure(ping_event));
-                self.identify
-                    .on_swarm_event(FromSwarm::DialFailure(identity_event));
-            }
-            FromSwarm::ListenFailure(e) => {
-                let (ping_handler, identity_handler) = e.handler.into_inner();
-                let ping_event = ListenFailure {
-                    handler: ping_handler,
-                    local_addr: e.local_addr,
-                    send_back_addr: e.send_back_addr,
-                };
-                let identity_event = ListenFailure {
-                    handler: identity_handler,
-                    local_addr: e.local_addr,
-                    send_back_addr: e.send_back_addr,
-                };
-                self.heartbeat
-                    .on_swarm_event(FromSwarm::ListenFailure(ping_event));
-                self.identify
-                    .on_swarm_event(FromSwarm::ListenFailure(identity_event));
-            }
-            FromSwarm::NewListener(e) => {
-                self.heartbeat.on_swarm_event(FromSwarm::NewListener(e));
-                self.identify.on_swarm_event(FromSwarm::NewListener(e));
-            }
-            FromSwarm::ExpiredListenAddr(e) => {
-                self.heartbeat
-                    .on_swarm_event(FromSwarm::ExpiredListenAddr(e));
-                self.identify
-                    .on_swarm_event(FromSwarm::ExpiredListenAddr(e));
-            }
-            FromSwarm::ListenerError(e) => {
-                self.heartbeat.on_swarm_event(FromSwarm::ListenerError(e));
-                self.identify.on_swarm_event(FromSwarm::ListenerError(e));
-            }
-            FromSwarm::ListenerClosed(e) => {
-                self.heartbeat.on_swarm_event(FromSwarm::ListenerClosed(e));
-                self.identify.on_swarm_event(FromSwarm::ListenerClosed(e));
-            }
-            FromSwarm::NewExternalAddr(e) => {
-                self.heartbeat.on_swarm_event(FromSwarm::NewExternalAddr(e));
-                self.identify.on_swarm_event(FromSwarm::NewExternalAddr(e));
-            }
-            FromSwarm::ExpiredExternalAddr(e) => {
-                self.heartbeat
-                    .on_swarm_event(FromSwarm::ExpiredExternalAddr(e));
-                self.identify
-                    .on_swarm_event(FromSwarm::ExpiredExternalAddr(e));
-            }
-            FromSwarm::NewListenAddr(e) => {
-                self.heartbeat.on_swarm_event(FromSwarm::NewListenAddr(e));
-                self.identify.on_swarm_event(FromSwarm::NewListenAddr(e));
-            }
-        }
-    }
-
-    fn poll(
-        &mut self,
-        cx: &mut Context<'_>,
-        params: &mut impl PollParameters,
-    ) -> Poll<NetworkBehaviourAction<Self::OutEvent, Self::ConnectionHandler>> {
-        if self.decay_interval.poll_tick(cx).is_ready() {
-            self.peer_manager.batch_update_score_with_decay();
-        }
-
-        if self.health_check.poll_tick(cx).is_ready() {
-            let disconnected_peers: Vec<_> = self
-                .peer_manager
-                .get_disconnected_reserved_peers()
-                .copied()
-                .collect();
-
-            for peer_id in disconnected_peers {
-                debug!(target: "fuel-p2p", "Trying to reconnect to reserved peer {:?}", peer_id);
-
-                self.peer_manager
-                    .pending_events
-                    .push_back(PeerInfoEvent::ReconnectToPeer(peer_id));
-            }
-        }
-
-        if let Some(event) = self.peer_manager.pending_events.pop_front() {
-            return Poll::Ready(NetworkBehaviourAction::GenerateEvent(event))
-        }
-
-        match self.heartbeat.poll(cx, params) {
-            Poll::Pending => {}
-            Poll::Ready(NetworkBehaviourAction::NotifyHandler {
-                peer_id,
-                handler,
-                event,
-            }) => {
-                return Poll::Ready(NetworkBehaviourAction::NotifyHandler {
-                    peer_id,
-                    handler,
-                    event: EitherOutput::First(event),
-                })
-            }
-            Poll::Ready(NetworkBehaviourAction::ReportObservedAddr {
-                address,
-                score,
-            }) => {
-                return Poll::Ready(NetworkBehaviourAction::ReportObservedAddr {
-                    address,
-                    score,
-                })
-            }
-            Poll::Ready(NetworkBehaviourAction::CloseConnection {
-                peer_id,
-                connection,
-            }) => {
-                return Poll::Ready(NetworkBehaviourAction::CloseConnection {
-                    peer_id,
-                    connection,
-                })
-            }
-            Poll::Ready(NetworkBehaviourAction::Dial { handler, opts }) => {
-                let handler =
-                    IntoConnectionHandler::select(handler, self.identify.new_handler());
-
-                return Poll::Ready(NetworkBehaviourAction::Dial { handler, opts })
-            }
-            Poll::Ready(NetworkBehaviourAction::GenerateEvent(HeartbeatEvent {
-                peer_id,
-                latest_block_height,
-            })) => {
-                let heartbeat_data = HeartbeatData::new(latest_block_height);
-
-                if let Some(previous_heartbeat) = self
-                    .get_peer_info(&peer_id)
-                    .and_then(|info| info.heartbeat_data.seconds_since_last_heartbeat())
-                {
-                    debug!(target: "fuel-p2p", "Previous hearbeat happened {:?} seconds ago", previous_heartbeat);
-                }
-
-                self.peer_manager.insert_peer_info(
-                    &peer_id,
-                    PeerInfoInsert::HeartbeatData(heartbeat_data),
-                );
-
-                let event = PeerInfoEvent::PeerInfoUpdated {
-                    peer_id,
-                    block_height: latest_block_height,
-                };
-                return Poll::Ready(NetworkBehaviourAction::GenerateEvent(event))
-            }
-        }
-
-        loop {
-            match self.identify.poll(cx, params) {
-                Poll::Pending => break,
-                Poll::Ready(NetworkBehaviourAction::NotifyHandler {
-                    peer_id,
-                    handler,
-                    event,
-                }) => {
-                    return Poll::Ready(NetworkBehaviourAction::NotifyHandler {
-                        peer_id,
-                        handler,
-                        event: EitherOutput::Second(event),
-                    })
-                }
-                Poll::Ready(NetworkBehaviourAction::ReportObservedAddr {
-                    address,
-                    score,
-                }) => {
-                    return Poll::Ready(NetworkBehaviourAction::ReportObservedAddr {
-                        address,
-                        score,
-                    })
-                }
-                Poll::Ready(NetworkBehaviourAction::CloseConnection {
-                    peer_id,
-                    connection,
-                }) => {
-                    return Poll::Ready(NetworkBehaviourAction::CloseConnection {
-                        peer_id,
-                        connection,
-                    })
-                }
-                Poll::Ready(NetworkBehaviourAction::Dial { handler, opts }) => {
-                    let handler = IntoConnectionHandler::select(
-                        self.heartbeat.new_handler(),
-                        handler,
-                    );
-                    return Poll::Ready(NetworkBehaviourAction::Dial { handler, opts })
-                }
-                Poll::Ready(NetworkBehaviourAction::GenerateEvent(event)) => {
-                    match event {
-                        IdentifyEvent::Received {
-                            peer_id,
-                            info:
-                                IdentifyInfo {
-                                    protocol_version,
-                                    agent_version,
-                                    mut listen_addrs,
-                                    ..
-                                },
-                        } => {
-                            if listen_addrs.len() > MAX_IDENTIFY_ADDRESSES {
-                                debug!(
-                                    target: "fuel-p2p",
-                                    "Node {:?} has reported more than {} addresses; it is identified by {:?} and {:?}",
-                                    peer_id, MAX_IDENTIFY_ADDRESSES, protocol_version, agent_version
-                                );
-                                listen_addrs.truncate(MAX_IDENTIFY_ADDRESSES);
-                            }
-
-                            self.peer_manager.insert_peer_info(
-                                &peer_id,
-                                PeerInfoInsert::ClientVersion(agent_version),
-                            );
-                            self.peer_manager.insert_peer_info(
-                                &peer_id,
-                                PeerInfoInsert::Addresses(listen_addrs.clone()),
-                            );
-
-                            let event = PeerInfoEvent::PeerIdentified {
-                                peer_id,
-                                addresses: listen_addrs,
-                            };
-                            return Poll::Ready(NetworkBehaviourAction::GenerateEvent(
-                                event,
-                            ))
-                        }
-                        IdentifyEvent::Error { peer_id, error } => {
-                            debug!(target: "fuel-p2p", "Identification with peer {:?} failed => {}", peer_id, error)
-                        }
-                        _ => {}
-                    }
-                }
-            }
-        }
-
-        Poll::Pending
-    }
-
-    fn on_connection_handler_event(
-        &mut self,
-        peer_id: PeerId,
-        connection_id: ConnectionId,
-        event: <<Self::ConnectionHandler as IntoConnectionHandler>::Handler as
-            ConnectionHandler>::OutEvent,
-    ) {
-        match event {
-            EitherOutput::First(heartbeat_event) => self
-                .heartbeat
-                .on_connection_handler_event(peer_id, connection_id, heartbeat_event),
-            EitherOutput::Second(identify_event) => self
-                .identify
-                .on_connection_handler_event(peer_id, connection_id, identify_event),
-        }
-    }
+#[derive(Debug, Clone, Copy)]
+pub struct PeerScoreUpdated {
+    pub should_ban: bool,
+    pub score: PeerScore,
 }
 
 // Info about a single Peer that we're connected to
@@ -597,17 +54,16 @@ impl Default for PeerInfo {
     }
 }
 
-enum PeerInfoInsert {
+pub enum PeerInfoInsert {
     Addresses(Vec<Multiaddr>),
     ClientVersion(String),
     HeartbeatData(HeartbeatData),
 }
 
 /// Manages Peers and their events
-#[derive(Debug, Default, Clone)]
-struct PeerManager {
+#[derive(Debug)]
+pub struct PeerManager {
     peer_score_config: PeerScoreConfig,
-    pending_events: VecDeque<PeerInfoEvent>,
     non_reserved_connected_peers: HashMap<PeerId, PeerInfo>,
     reserved_connected_peers: HashMap<PeerId, PeerInfo>,
     reserved_peers: HashSet<PeerId>,
@@ -616,7 +72,7 @@ struct PeerManager {
 }
 
 impl PeerManager {
-    fn new(
+    pub fn new(
         peer_score_config: PeerScoreConfig,
         reserved_peers: HashSet<PeerId>,
         connection_state: Arc<RwLock<ConnectionState>>,
@@ -624,7 +80,6 @@ impl PeerManager {
     ) -> Self {
         Self {
             peer_score_config,
-            pending_events: VecDeque::default(),
             non_reserved_connected_peers: HashMap::with_capacity(max_non_reserved_peers),
             reserved_connected_peers: HashMap::with_capacity(reserved_peers.len()),
             reserved_peers,
@@ -633,53 +88,61 @@ impl PeerManager {
         }
     }
 
-    fn batch_update_score_with_decay(&mut self) {
-        for (_, peer_info) in &mut self.non_reserved_connected_peers {
-            peer_info.score = peer_info.score * DECAY_PEER_SCORE;
+    pub fn should_ban_peer_for_gossip(
+        &self,
+        peer_id: &PeerId,
+        gossip_score: f64,
+    ) -> bool {
+        self.non_reserved_connected_peers.contains_key(peer_id)
+            && gossip_score < self.peer_score_config.get_min_gossipsub_score()
+    }
+
+    pub fn batch_update_score_with_decay(&mut self) {
+        for peer_info in self.non_reserved_connected_peers.values_mut() {
+            peer_info.score *= DECAY_PEER_SCORE;
         }
     }
 
-    fn update_peer_score_with(
+    pub fn update_peer_score_with(
         &mut self,
         peer_id: PeerId,
         score: PeerScore,
-    ) -> Option<PeerScore> {
+    ) -> Option<PeerScoreUpdated> {
         if let Some(peer) = self.non_reserved_connected_peers.get_mut(&peer_id) {
             // score should not go over `MAX_PEER_SCORE`
-            let new_score = peer.score + score;
-            peer.score = MAX_PEER_SCORE.min(new_score);
+            let new_score = MAX_PEER_SCORE.min(peer.score + score);
+            peer.score = new_score;
 
-            if peer.score < self.peer_score_config.get_min_app_score() {
-                // ban the peer with low score
-                self.pending_events
-                    .push_front(PeerInfoEvent::BanPeer { peer_id })
-            }
+            let should_ban = new_score < self.peer_score_config.get_min_app_score();
 
-            Some(peer.score)
+            Some(PeerScoreUpdated {
+                should_ban,
+                score: new_score,
+            })
         } else {
             log_missing_peer(&peer_id);
             None
         }
     }
 
-    fn total_peers_connected(&self) -> usize {
+    pub fn total_peers_connected(&self) -> usize {
         self.reserved_connected_peers.len() + self.non_reserved_connected_peers.len()
     }
 
-    fn get_peers_ids(&self) -> impl Iterator<Item = &PeerId> {
+    pub fn get_peers_ids(&self) -> impl Iterator<Item = &PeerId> {
         self.non_reserved_connected_peers
             .keys()
             .chain(self.reserved_connected_peers.keys())
     }
 
-    fn get_peer_info(&self, peer_id: &PeerId) -> Option<&PeerInfo> {
+    pub fn get_peer_info(&self, peer_id: &PeerId) -> Option<&PeerInfo> {
         if self.reserved_peers.contains(peer_id) {
             return self.reserved_connected_peers.get(peer_id)
         }
         self.non_reserved_connected_peers.get(peer_id)
     }
 
-    fn insert_peer_info(&mut self, peer_id: &PeerId, data: PeerInfoInsert) {
+    pub fn insert_peer_info(&mut self, peer_id: &PeerId, data: PeerInfoInsert) {
         let peers = if self.reserved_peers.contains(peer_id) {
             &mut self.reserved_connected_peers
         } else {
@@ -698,29 +161,26 @@ impl PeerManager {
         }
     }
 
-    fn get_disconnected_reserved_peers(&self) -> impl Iterator<Item = &PeerId> {
+    pub fn get_disconnected_reserved_peers(&self) -> impl Iterator<Item = &PeerId> {
         self.reserved_peers
             .iter()
             .filter(|peer_id| !self.reserved_connected_peers.contains_key(peer_id))
     }
 
     /// Handles the first connnection established with a Peer
-    fn handle_initial_connection(&mut self, peer_id: PeerId) {
-        let non_reserved_peers_connected = self.non_reserved_connected_peers.len();
-
+    /// Returns `true` signaling that the peer should be disconnected
+    pub fn handle_initial_connection(
+        &mut self,
+        peer_id: PeerId,
+        addresses: Vec<Multiaddr>,
+    ) -> bool {
         // if the connected Peer is not from the reserved peers
         if !self.reserved_peers.contains(&peer_id) {
+            let non_reserved_peers_connected = self.non_reserved_connected_peers.len();
             // check if all the slots are already taken
             if non_reserved_peers_connected >= self.max_non_reserved_peers {
-                // Too many peers already connected, disconnect the Peer with the first priority.
-                self.pending_events.push_front(PeerInfoEvent::TooManyPeers {
-                    peer_to_disconnect: peer_id,
-                });
-
-                // early exit, we don't want to report new peer connection
-                // nor insert it in `connected_peers`
-                // since we're going to disconnect it anyways
-                return
+                // Too many peers already connected, disconnect the Peer
+                return true
             }
 
             if non_reserved_peers_connected + 1 == self.max_non_reserved_peers {
@@ -737,23 +197,24 @@ impl PeerManager {
                 .insert(peer_id, PeerInfo::default());
         }
 
-        self.pending_events
-            .push_back(PeerInfoEvent::PeerConnected(peer_id));
+        self.insert_peer_info(&peer_id, PeerInfoInsert::Addresses(addresses));
+
+        false
     }
 
     /// Handles on peer's last connection getting disconnected
-    fn handle_peer_disconnect(&mut self, peer_id: PeerId) {
+    /// Returns 'true' signaling we should try reconnecting
+    pub fn handle_peer_disconnect(&mut self, peer_id: PeerId) -> bool {
         // try immediate reconnect if it's a reserved peer
         let is_reserved = self.reserved_peers.contains(&peer_id);
-        let is_removed;
 
         if !is_reserved {
-            is_removed = self.non_reserved_connected_peers.remove(&peer_id).is_some();
+            // check were all the slots taken prior to this disconnect
+            let all_slots_taken = self.max_non_reserved_peers
+                == self.non_reserved_connected_peers.len() + 1;
 
-            // check were all the slots full prior to this disconnect
-            if is_removed
-                && self.max_non_reserved_peers
-                    == self.non_reserved_connected_peers.len() + 1
+            if self.non_reserved_connected_peers.remove(&peer_id).is_some()
+                && all_slots_taken
             {
                 // since all the slots were full prior to this disconnect
                 // let's allow new peer non-reserved peers connections
@@ -761,21 +222,15 @@ impl PeerManager {
                     connection_state.allow_new_peers();
                 }
             }
-        } else {
-            is_removed = self.reserved_connected_peers.remove(&peer_id).is_some();
-        }
 
-        if is_removed {
-            self.pending_events
-                .push_back(PeerInfoEvent::PeerDisconnected {
-                    peer_id,
-                    should_reconnect: is_reserved,
-                })
+            false
+        } else {
+            self.reserved_connected_peers.remove(&peer_id).is_some()
         }
     }
 
     /// Find a peer that is holding the given block height.
-    fn get_peer_id_with_height(&self, height: &BlockHeight) -> Option<PeerId> {
+    pub fn get_peer_id_with_height(&self, height: &BlockHeight) -> Option<PeerId> {
         let mut range = rand::thread_rng();
         // TODO: Optimize the selection of the peer.
         //  We can store pair `(peer id, height)` for all nodes(reserved and not) in the
@@ -792,7 +247,7 @@ impl PeerManager {
 }
 
 #[derive(Debug, Default, Clone, Copy)]
-pub(crate) struct ConnectionState {
+pub struct ConnectionState {
     peers_allowed: bool,
 }
 
@@ -803,7 +258,7 @@ pub struct HeartbeatData {
 }
 
 impl HeartbeatData {
-    fn new(block_height: BlockHeight) -> Self {
+    pub fn new(block_height: BlockHeight) -> Self {
         Self {
             block_height: Some(block_height),
             last_heartbeat: Some(Instant::now()),
@@ -908,7 +363,7 @@ mod tests {
 
         // try connecting all the random peers
         for peer_id in &random_peers {
-            peer_manager.handle_initial_connection(*peer_id);
+            peer_manager.handle_initial_connection(*peer_id, vec![]);
         }
 
         assert_eq!(peer_manager.total_peers_connected(), max_non_reserved_peers);
@@ -923,7 +378,7 @@ mod tests {
 
         // try connecting all the reserved peers
         for peer_id in &reserved_peers {
-            peer_manager.handle_initial_connection(*peer_id);
+            peer_manager.handle_initial_connection(*peer_id, vec![]);
         }
 
         assert_eq!(peer_manager.total_peers_connected(), reserved_peers.len());
@@ -931,7 +386,7 @@ mod tests {
         // try connecting random peers
         let random_peers = get_random_peers(10);
         for peer_id in &random_peers {
-            peer_manager.handle_initial_connection(*peer_id);
+            peer_manager.handle_initial_connection(*peer_id, vec![]);
         }
 
         // the number should stay the same
@@ -947,7 +402,7 @@ mod tests {
 
         // try connecting all the reserved peers
         for peer_id in &reserved_peers {
-            peer_manager.handle_initial_connection(*peer_id);
+            peer_manager.handle_initial_connection(*peer_id, vec![]);
         }
 
         // disconnect a single reserved peer
@@ -956,7 +411,7 @@ mod tests {
         // try connecting random peers
         let random_peers = get_random_peers(max_non_reserved_peers * 2);
         for peer_id in &random_peers {
-            peer_manager.handle_initial_connection(*peer_id);
+            peer_manager.handle_initial_connection(*peer_id, vec![]);
         }
 
         // there should be an available slot for a reserved peer
@@ -966,7 +421,7 @@ mod tests {
         );
 
         // reconnect the disconnected reserved peer
-        peer_manager.handle_initial_connection(*reserved_peers.first().unwrap());
+        peer_manager.handle_initial_connection(*reserved_peers.first().unwrap(), vec![]);
 
         // all the slots should be taken now
         assert_eq!(
